Test proposed by Srivastava and Fujikoshi (2006)

Srivastava and Fujikoshi (2006)'s test for general linear hypothesis testing (GLHT) problem for high-dimensional data with assuming that underlying covariance matrices are the same.

Usage

glht_sf2006(Y,X,C)

Arguments

Y

An \(n\times p\) response matrix obtained by independently observing a \(p\)-dimensional response variable for \(n\) subjects.

X

A known \(n\times k\) full-rank design matrix with \(\operatorname{rank}(\boldsymbol{G})=k<n\).

C

A known matrix of size \(q\times k\) with \(\operatorname{rank}(\boldsymbol{C})=q<k\).

Value

A (list) object of S3 class htest containing the following elements:

statistic

the test statistic proposed by Srivastava and Fujikoshi (2006).

p.value

the \(p\)-value of the test proposed by Srivastava and Fujikoshi (2006).

Details

A high-dimensional linear regression model can be expressed as $$\boldsymbol{Y}=\boldsymbol{X\Theta}+\boldsymbol{\epsilon},$$ where \(\Theta\) is a \(k\times p\) unknown parameter matrix and \(\boldsymbol{\epsilon}\) is an \(n\times p\) error matrix.

It is of interest to test the following GLHT problem $$H_0: \boldsymbol{C\Theta}=\boldsymbol{0}, \quad \text { vs. } \quad H_1: \boldsymbol{C\Theta} \neq \boldsymbol{0}.$$

Srivastava and Fujikoshi (2006) proposed the following test statistic: $$T_{SF}=\left[2q\hat{a}_2(1+n^{-1}q)\right]^{-1/2}\left[\frac{\operatorname{tr}(\boldsymbol{B})}{\sqrt{p}}-\frac{q}{\sqrt{n}}\frac{\operatorname{tr}(\boldsymbol{W})}{\sqrt{np}}\right].$$ where \(\boldsymbol{W}\) and \(\boldsymbol{B}\) are the matrix of sum of squares and products due to error and the error, respectively, and \(\hat{a}_2=[\operatorname{tr}(\boldsymbol{W}^2)-\operatorname{tr}^2(\boldsymbol{W})/n]/[(n-1)(n+2)p]\). They showed that under the null hypothesis, \(T_{SF}\) is asymptotically normally distributed.

References

Srivastava MS, Fujikoshi Y (2006). “Multivariate analysis of variance with fewer observations than the dimension.” Journal of Multivariate Analysis, 97(9), 1927--1940. doi:10.1016/j.jmva.2005.08.010 .

Examples

set.seed(1234)
k <- 3
q <- k-1
p <- 50
n <- c(25,30,40)
rho <- 0.01
Theta <- matrix(rep(0,k*p),nrow=k)
X <- matrix(c(rep(1,n[1]),rep(0,sum(n)),rep(1,n[2]),rep(0,sum(n)),rep(1,n[3])),ncol=k,nrow=sum(n))
y <- (-2*sqrt(1-rho)+sqrt(4*(1-rho)+4*p*rho))/(2*p)
x <- y+sqrt((1-rho))
Gamma <- matrix(rep(y,p*p),nrow=p)
diag(Gamma) <- rep(x,p)
U <- matrix(ncol = sum(n),nrow=p)
for(i in 1:sum(n)){
U[,i] <- rnorm(p,0,1)
}
Y <- X%*%Theta+t(U)%*%Gamma
C <- cbind(diag(q),-rep(1,q))
glht_sf2006(Y,X,C)
#> 
#> 
#> 
#> data:  
#> statistic = 0.56836, p-value = 0.2849
#>